Start a bodies index
I want a static list of bodies so we can start doing search
and provide an index. The bodies.edn file doesn't have all
the major bodies supported by HORIZONS, but we'll get there.

This commit makes the bodies file, and makes space for it
in the app.

NOTE: Since liberator can't serialize a PersistentHashSet,
requesting the bodies index will fail. I'm working on it.